+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 All notable changes to the Viewer Metrics Chrome Extension will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [0.9.7] - 2025-11-12
+Minor fixes
+Added ability to change user from tracking page
+Improved viewer details popup
 
-## [0.9.5] - 2025-10-22
+## [0.9.6] - 2025-11-12
 
 ### 🎉 Pre-Release